Fixed a bug where deleting a section, global set, or tag set wouldn't delete its field layout in the database

+++ src/services/TagsService.php
@@ -214,8 +214,37 @@ class TagsService extends BaseApplicationComponent
 	*/
 	public function deleteTagSetById($tagSetId)
 	{
-		$affectedRows = craft()->db->createCommand()->delete('tagsets', array('id' => $tagSetId));
-		return (bool) $affectedRows;
+		if (!$tagSetId)
+		{
+			return false;
+		}
+
+		$transaction = craft()->db->beginTransaction();
+		try
+		{
+			// Delete the field layout
+			$fieldLayoutId = craft()->db->createCommand()
+				->select('fieldLayoutId')
+				->from('tagsets')
+				->where(array('id' => $tagSetId))
+				->queryScalar();
+
+			if ($fieldLayoutId)
+			{
+				craft()->fields->deleteLayoutById($fieldLayoutId);
+			}
+
+			$affectedRows = craft()->db->createCommand()->delete('tagsets', array('id' => $tagSetId));
+
+			$transaction->commit();
+
+			return (bool) $affectedRows;
+		}
+		catch (\Exception $e)
+		{
+			$transaction->rollBack();
+			throw $e;
+		}
 	}
 
 	// Tags
